Question community link (#50)
* Display community link on questions

Displays a link with the name of the community that a question was posted in. Displays normally if the question was not posted in a community.

* Tests for addQuestionToCommunityModel

Added missing tests for addQuestionToCommunityModel.

* updated styling

updating styling of the questionview component


* Updated AddQuestionToCommunity. Added real time updates.

AddQuestionToCommunity now returns the question rather than the updated community. Added real time updated for community names showing up on questions posted and questions posted in communities.